pyproject.toml
[project]
name = "mcp-example"
version = "0.1.0"
description = "A MCP Server that allows"
readme = "README.md"
requires-python = ">=3.13.1"
dependencies = ["mcp[cli]>=1.3.0", "python-dotenv>=1.0.1"]

[dependency-groups]
dev = ["ruff>=0.9.9"]
